refactor(frontend): extract rendering/ack helpers into rendering.ts
dataAgeClass and the command-ack/server-misalignment state machine
(commandAckDisplay, commandAckOutcome, supersededRtlInEffect,
assetServerMisalignment, and the staleness/ack-timeout threshold
constants) lived as unexported consts inside fss-main-page.tsx, so they
could only be exercised by rendering the component. Move them to
rendering.ts, a plain module with no React/DOM dependency, and import
them back into fss-main-page.tsx for JSX use. No behaviour change;
enables direct unit testing (see the following commits).
